adapt suggestions
