+++ pkg/client/kubernetes/client.go
@@ -28,7 +28,6 @@ import (
 	securityinstall "github.com/gardener/gardener/pkg/apis/security/install"
 	seedmanagementinstall "github.com/gardener/gardener/pkg/apis/seedmanagement/install"
 	settingsinstall "github.com/gardener/gardener/pkg/apis/settings/install"
-	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
 )
 
 const (
@@ -344,7 +343,7 @@ func newClientSet(conf *Config) (Interface, error) {
 		restClient: kubernetes.Discovery().RESTClient(),
 
 		applier:     NewApplier(runtimeClient, conf.clientOptions.Mapper),
-		podExecutor: kubernetesutils.NewPodExecutor(conf.restConfig),
+		podExecutor: NewPodExecutor(conf.restConfig),
 
 		client:    runtimeClient,
 		apiReader: runtimeAPIReader,
