+++ pkg/client/kubernetes/fake/builder.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/gardener/gardener/pkg/chartrenderer"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
-	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
 )
 
 // ClientSetBuilder is a builder for fake ClientSets
@@ -20,7 +19,7 @@ type ClientSetBuilder struct {
 	applier       kubernetes.Applier
 	chartRenderer chartrenderer.Interface
 	chartApplier  kubernetes.ChartApplier
-	podExecutor   kubernetesutils.PodExecutor
+	podExecutor   kubernetes.PodExecutor
 	restConfig    *rest.Config
 	client        client.Client
 	apiReader     client.Reader
@@ -54,7 +53,7 @@ func (b *ClientSetBuilder) WithChartApplier(chartApplier kubernetes.ChartApplier
 }
 
 // WithPodExecutor sets the podExecutor attribute of the builder.
-func (b *ClientSetBuilder) WithPodExecutor(podExecutor kubernetesutils.PodExecutor) *ClientSetBuilder {
+func (b *ClientSetBuilder) WithPodExecutor(podExecutor kubernetes.PodExecutor) *ClientSetBuilder {
 	b.podExecutor = podExecutor
 	return b
 }
