+++ pkg/client/kubernetes/fake/clientset.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/gardener/gardener/pkg/chartrenderer"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
-	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
 )
 
 var _ kubernetes.Interface = &ClientSet{}
@@ -25,7 +24,7 @@ type ClientSet struct {
 	applier       kubernetes.Applier
 	chartRenderer chartrenderer.Interface
 	chartApplier  kubernetes.ChartApplier
-	podExecutor   kubernetesutils.PodExecutor
+	podExecutor   kubernetes.PodExecutor
 	restConfig    *rest.Config
 	client        client.Client
 	apiReader     client.Reader
@@ -56,7 +55,7 @@ func (c *ClientSet) ChartApplier() kubernetes.ChartApplier {
 }
 
 // PodExecutor returns a PodExecutor.
-func (c *ClientSet) PodExecutor() kubernetesutils.PodExecutor {
+func (c *ClientSet) PodExecutor() kubernetes.PodExecutor {
 	return c.podExecutor
 }
